+++ apps/mobile/src/app/(tabs)/settings/index.tsx
@@ -15,6 +15,7 @@ export default function SettingsPage() {
 	const router = useRouter()
 	const account = useAppStore((state) => state.bbplayerAccount)
 	const hasBilibiliCookie = useAppStore((state) => state.hasBilibiliCookie())
+	const bilibiliUserInfo = useAppStore((state) => state.bilibiliUserInfo)
 
 	return (
 		<View style={[styles.container, { backgroundColor: colors.background }]}>
@@ -40,6 +41,8 @@ export default function SettingsPage() {
 						{ paddingBottom: insets.bottom + (haveTrack ? 132 : 40) },
 					]}
 					contentInsetAdjustmentBehavior='automatic'
+					showsVerticalScrollIndicator
+					persistentScrollbar // 我看哪个 b 还说看不见这是可滚动的？！
 				>
 					<List.Item
 						title='外观'
@@ -114,32 +117,10 @@ export default function SettingsPage() {
 					/>
 					<Divider style={styles.divider} />
 					<List.Item
-						title='BBPlayer 账号'
-						description={
-							account
-								? `${account.name}（@${account.username}）`
-								: '注册、登录、个人资料'
-						}
-						left={(props) => (
-							<List.Icon
-								{...props}
-								icon='account-circle'
-							/>
-						)}
-						right={(props) => (
-							<List.Icon
-								{...props}
-								icon='chevron-right'
-							/>
-						)}
-						onPress={() => router.push('/settings/account' as never)}
-					/>
-					<Divider style={styles.divider} />
-					<List.Item
 						title='Bilibili 账号'
 						description={
-							hasBilibiliCookie
-								? '已登录，管理 Cookie 和播放记录'
+							hasBilibiliCookie && bilibiliUserInfo
+								? `${bilibiliUserInfo.name} ( uid${bilibiliUserInfo.mid} )`
 								: '扫码、手机号或 Cookie 登录'
 						}
 						left={(props) => (
@@ -155,6 +136,28 @@ export default function SettingsPage() {
 							/>
 						)}
 						onPress={() => router.push('/settings/bilibili-account' as never)}
+					/>
+					<Divider style={styles.divider} />
+					<List.Item
+						title='BBPlayer 账号'
+						description={
+							account
+								? `${account.name} ( @${account.username} )`
+								: '注册、登录、个人资料'
+						}
+						left={(props) => (
+							<List.Icon
+								{...props}
+								icon='account-circle'
+							/>
+						)}
+						right={(props) => (
+							<List.Icon
+								{...props}
+								icon='chevron-right'
+							/>
+						)}
+						onPress={() => router.push('/settings/account' as never)}
 					/>
 					<Divider style={styles.divider} />
 					<List.Item
@@ -196,7 +199,7 @@ export default function SettingsPage() {
 					<Divider style={styles.divider} />
 					<List.Item
 						title='关于 BBPlayer'
-						description='版本、链接、开源许可证'
+						description='版本、开源许可证'
 						left={(props) => (
 							<List.Icon
 								{...props}